Separating dialog helper methods
